Confirmación para eliminación de tarea en vista detalle

+++ src/app/item-detail/item-detail.component.ts
@@ -12,10 +12,22 @@ import { areEqualTasks as equal, DEFAULT_TASK } from '../utils/task-utils';
 export class ItemDetailComponent implements OnInit {
   originalTask: TaskItem = DEFAULT_TASK;
   task: TaskItem = DEFAULT_TASK;
+  private confirmDelete = false;
+  private timeout: any;
 
   constructor(private router: Router,
               private route: ActivatedRoute,
               private taskService: TaskService) {
+  }
+
+  private static toggleButtonStatus(button: HTMLButtonElement): void {
+    if (button.innerText === 'Eliminar') {
+      button.innerText = 'Confirmar';
+      button.classList.replace('btn-danger', 'btn-warning');
+    } else {
+      button.innerText = 'Eliminar';
+      button.classList.replace('btn-warning', 'btn-danger');
+    }
   }
 
   saveChanges(): void {
@@ -47,9 +59,21 @@ export class ItemDetailComponent implements OnInit {
     this.router.navigate(['/']);
   }
 
-  deleteTask(): void {
-    this.taskService.removeTask(this.task.id);
-    this.router.navigate(['/']);
+  deleteTask(button: HTMLButtonElement): void {
+    if (this.confirmDelete) {
+      clearInterval(this.timeout);
+
+      this.taskService.removeTask(this.task.id);
+      this.router.navigate(['/']);
+    } else {
+      ItemDetailComponent.toggleButtonStatus(button);
+      this.confirmDelete = true;
+
+      this.timeout = setTimeout(() => {
+        ItemDetailComponent.toggleButtonStatus(button);
+        this.confirmDelete = false;
+      }, 2500);
+    }
   }
 
   ngOnInit(): void {
